Add setGameOver() function
The first two lines disable the form text input and button by setting their disabled properties to true. This is necessary, because if we didn't, the user could submit more guesses after the game is over, which would mess things up.
The next three lines generate a new <button> element, set its text label to "Start new game", and add it to the bottom of our existing HTML.

The final line sets an event listener on our new button so that when it is clicked, a function called resetGame() is run.