Add resetGame() function
This rather long block of code completely resets everything to how it was at the start of the game, so the player can have another go. It:

Puts the guessCount back down to 1.

Empties all the text out of the information paragraphs. We select all paragraphs inside <div class="resultParas"></div>, then loop through each one, setting their textContent to '' (an empty string).

Removes the reset button from our code.

Enables the form elements, and empties and focuses the text field, ready for a new guess to be entered.

Removes the background color from the lastResult paragraph.

Generates a new random number so that you are not just guessing the same number again!

At this point, you should have a fully working (simple) game — congratulations!